Create Factorial.dart
We define a function called calculateFactorial that takes an integer n as an argument.

In the calculateFactorial function, we use recursion. If n is equal to 0, we return 1 (the factorial of 0 is 1). Otherwise, we multiply n by the result of `calculateFactorial(n - 1). This is what makes the algorithm recursive, as it calls itself with a reduced value in each iteration.

In the main function, we define a number (in this case, 5) for which we want to calculate the factorial. You can change this number as needed.

Then, we call the calculateFactorial function with the desired number and store the result in the factorial variable.

Finally, we print the result using print to display the factorial of the provided number.